+++ packages/service/test/fixtures/bad-typescript-plugin/platformatic.service.json
@@ -5,7 +5,7 @@
     },
     "hostname": "127.0.0.1",
     "port": "3042",
-    "pluginTimeout": 30000,
+    "pluginTimeout": 60000,
     "keepAliveTimeout": 1
   },
   "plugins": {
