+++ packages/service/test/fixtures/typescript-autoload/tsconfig.json
@@ -2,10 +2,10 @@
   "compilerOptions": {
     "module": "commonjs",
     "esModuleInterop": true,
-    "target": "es2019",
+    "target": "es2020",
     "moduleResolution": "node",
-    "sourceMap": true,
-    "pretty": true,
+    "sourceMap": false,
+    "pretty": false,
     "noEmitOnError": true,
     "outDir": "dist"
   },
