+++ packages/service/test/fixtures/typescript-plugin-custom-flags/tsconfig.json
@@ -2,10 +2,10 @@
   "compilerOptions": {
     "module": "commonjs",
     "esModuleInterop": true,
-    "target": "es6",
+    "target": "es2020",
     "moduleResolution": "node",
-    "sourceMap": true,
-    "pretty": true,
+    "sourceMap": false,
+    "pretty": false,
     "noEmitOnError": true,
     "outDir": "dist"
   },
